services/stt/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.0
torchaudio==2.1.0
openai-whisper==20231117
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1